Move decorator to api_auth
